rename functions directory and files to basic
